requirements.txt
qiskit==0.45.2
qiskit_terra==0.45.2
tqdm==4.65.0
python-dotenv==0.19.2